+++ src/App.js
@@ -8,8 +8,8 @@ import LostWorld from './containers/LostWorld/LostWorld';
 import Logout from './containers/Auth/Logout/Logout'
 import './App.module.css';
 import Profile from './components/Profile/Profile';
-import Form from './components/Test/test';
 import * as actions from './store/actions/index';
+import Test from "./components/Test/Test";
 
 const App = props => {
 
@@ -23,7 +23,7 @@ const App = props => {
             <Route path="/login" component={Login}/>
             <Route path="/register" component={Register}/>
             <Route path="/myProfile" component={Profile}/>
-            <Route path="/test" component={Form}/>
+            <Route path="/test/:id" component={Test}/>
             <Route path="/" exact component={LostWorld}/>
         </Switch>
     );
@@ -35,7 +35,7 @@ const App = props => {
                 <Route exact path="/logout" component={Logout}/>
                 <Route path="/register" component={Register}/>
                 <Route path="/myProfile" component={Profile}/>
-                <Route path="/test" component={Form}/>
+                <Route path="/test/:id" component={Test}/>
                 <Route path="/" exact component={LostWorld}/>
             </Switch>
         )
